// src/app/Interfaces/blog_interface.ts
export interface IBlog {
  BlogDescription: string;
  BlogImageURL: string|null|ArrayBuffer;
  BlogImageURL1: string;
  BlogImageURL2: string;
  BlogImageURL3: string;
  BlogImageURL4: string;
  BlogMetaDataDescription: string;
  BlogMetaDataKeywords: string;
  BlogMetaDataTitle: string;
  BlogAltText: string;
  BlogName: string;
  BlogNewURL: string;
  BlogOldURL: string;
  BlogSchema: string;
  Featured_Blog_Flag: string;
  OrganiztionId: string;
  WebSiteName: string;
  id: string;
  DelFlag: string;

}